requirements.txt
dlib==19.8.1
matplotlib==2.2.2
numpy==1.16.4
opencv-contrib-python==4.1.0.25
tensorflow-gpu==2.0.0
tqdm==4.23.4